fix: use index-based position tracking in YAML parser
Addresses review feedback on #17392:

1. Replaced lines.indexOf(line) with explicit index tracking (for-let
   instead of for-of) in parseSimpleYaml to handle duplicate key lines
   correctly when peeking ahead to distinguish arrays from objects.

2. Extracted shared parse-skill-yaml.mjs module used by both
   generate-catalog.mjs and lint-skill-spec.mjs, eliminating parser
   divergence where the lint parser lacked array/inline-JSON support.

Closes #17392.
